requirements.txt
feedparser==5.2.1
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
py-dateutil==2.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
Werkzeug==0.11.3
wheel==0.26.0
